[Codegen 120] Refactor: Extract findComponentConfig(...) to parsers-commons.js
